Added useRouter from expo-router for navigation.
Changed setfirst to setIsSubmitting for clarity.

Removed the unused submit function and integrated its functionality into handleVerify.

Added auto-focus functionality to move to the next input when a digit is entered.

Implemented a simulated API call in handleVerify with navigation to the next screen.

Adjusted the styling of the CustomButton to make it consistent with the layout.

Renamed text3 to disclaimer in the styles for better clarity.

Adjusted the font size of the external links to make them more proportional to the disclaimer text.

Added useRef to create references for the TextInput components, allowing for programmatic focus.